HTCONDOR-1768 SetAttributeInt() now takes a long long value
This allows proper handling of time_t values.
These were spotted by Coverity.

+++ src/condor_schedd.V6/dedicated_scheduler.cpp
@@ -744,7 +744,7 @@ DedicatedScheddNegotiate::scheduler_handleJobRejected(PROC_ID job_id,char const 
 
 	SetAttributeInt(
 		job_id.cluster, job_id.proc,
-		ATTR_LAST_REJ_MATCH_TIME, (int)time(0), NONDURABLE);
+		ATTR_LAST_REJ_MATCH_TIME, time(0), NONDURABLE);
 }
 
 void
